+++ srcs/execution/exec_pipe_list.c
@@ -6,7 +6,7 @@
 /*   By: jnakahod <jnakahod@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/01 21:29:47 by jnakahod          #+#    #+#             */
-/*   Updated: 2021/10/18 11:35:34 by jnakahod         ###   ########.fr       */
+/*   Updated: 2021/10/18 14:27:42 by jnakahod         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,60 +62,6 @@ void	exec_buildin(t_pipe_list *pipe_list, t_builtin_id builtin_id)
 
 	init_builtin_funcs(builtin_funcs);
   builtin_funcs[builtin_id](pipe_list->cmd_args, FALSE);
-}
-
-t_bool	has_heredoc(t_heredoc_to_fd *heredoc)
-{
-	if (heredoc)
-		return (TRUE);
-	else
-		return (FALSE);
-}
-
-char	*get_tmp_file_path(int	i)
-{
-	char	*num;
-	char	*file_path;
-
-	num = ft_itoa(i);
-	if (!*num)
-		return (NULL);
-	file_path = ft_strjoin(TMPFILE, num);
-	free(num);
-	return (file_path);
-}
-
-int	heredoc_to_fd(t_heredoc_to_fd *heredoc)
-{
-	int		tmp_file_fd;
-	char	*contents;
-	static int i = -1;
-
-	heredoc->tmp_file_path = get_tmp_file_path(++i);
-	if (!heredoc->tmp_file_path)
-		return (-1);
-	tmp_file_fd = open(heredoc->tmp_file_path, O_RDWR | O_CREAT | O_EXCL, 0666);
-	if (tmp_file_fd < 0)
-		return (-1);
-	contents = heredoc->contents;
-	if (write(tmp_file_fd, contents, ft_strlen(contents)) < 0)
-	{
-		close(tmp_file_fd);
-		unlink(heredoc->tmp_file_path);
-		return (-1);
-	}
-	return (tmp_file_fd);
-}
-
-t_result write_heredoc(t_heredoc_to_fd *heredoc)
-{
-	if (!has_heredoc(heredoc))
-		return (SUCCESS);
-	heredoc->tmp_fd = heredoc_to_fd(heredoc);
-	if (heredoc->tmp_fd < 0)
-		return (FAILURE);
-	else
-		return (SUCCESS);
 }
 
 void	exec_simple_cmd(t_pipe_list *pipe_list)
